requirements.txt
asgiref==3.4.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
cryptography==35.0.0
defusedxml==0.7.1
Django==3.2.9
django-cors-headers
django-allauth==0.46.0
django-crispy-forms==1.13.0
django-environ==0.8.1
django-mysql==4.2.0
dj_rest_auth
django-rest-auth==0.9.5
djangorestframework<=3.11.0
djangorestframework_simplejwt
djangorestframework-jwt
entrypoints==0.3
environ==1.0
gunicorn==20.1.0
idna==3.3
mysqlclient==2.0.3
oauthlib==3.1.1
pycparser==2.21
PyJWT==2.3.0
python3-openid==3.2.0
pytz==2021.3
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
social-auth-core==4.1.0
sqlparse==0.4.2
urllib3==1.26.7
whitenoise==5.3.0
